security commit

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -1,9 +1,12 @@
 package net.engineeringdigest.journalApp.service;
 
+import lombok.extern.slf4j.Slf4j;
 import net.engineeringdigest.journalApp.entity.JournalEntry;
 import net.engineeringdigest.journalApp.entity.User;
 import net.engineeringdigest.journalApp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,6 +16,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Component
+@Slf4j
 public class JournalEntryService {
 
     @Autowired
@@ -20,6 +24,9 @@ public class JournalEntryService {
 
     @Autowired
     private UserService userService;
+
+    private static final Logger logger = LoggerFactory.getLogger(JournalEntryService.class);
+
     @Transactional
     public void saveEntry(JournalEntry journalEntry, String userName){
         try{
@@ -28,7 +35,7 @@ public class JournalEntryService {
             JournalEntry saved =  journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
             //user.setUsername(null);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         }catch(Exception e){
             System.out.println(e);
             throw new RuntimeException("An error occured while saving the journal entry", e);
@@ -43,11 +50,22 @@ public class JournalEntryService {
     public Optional<JournalEntry> findById(ObjectId id){
         return journalEntryRepository.findById(id);
     }
-    public void deleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed = false;
+        try{
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }catch (Exception e) {
+            //System.out.println(e);
+            log.error("Error Occured while deleting the journal entry", e);
+            throw new RuntimeException("An error occured while deleting the journal entry", e);
+        }
+        return removed;
     }
 }
